# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/staff/auth/login": {
      "post": {
        "operationId": "StaffAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffSessionDto"
                }
              }
            }
          }
        },
        "tags": [
          "staff-auth"
        ]
      }
    },
...
    "/staff/auth/logout": {
      "post": {
        "operationId": "StaffAuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-auth"
        ]
      }
    },
    "/staff/me": {
      "get": {
        "operationId": "StaffMeController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffMeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-auth"
        ]
      }
    },
...
    "/services/{id}": {
      "get": {
        "operationId": "ServiceController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicServiceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "marketplace"
        ]
      }
    },
    "/seller/services": {
      "post": {
        "operationId": "SellerServiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      },
      "get": {
        "operationId": "SellerServiceController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      }
    },
    "/seller/services/{id}": {
      "get": {
        "operationId": "SellerServiceController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      },
      "patch": {
        "operationId": "SellerServiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      }
    },
    "/seller/services/{id}/submit": {
      "post": {
        "operationId": "SellerServiceController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      }
    },
    "/seller/services/{id}/pause": {
      "post": {
        "operationId": "SellerServiceController_pause",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      }
    },
    "/seller/services/{id}/resume": {
      "post": {
        "operationId": "SellerServiceController_resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      }
    },
    "/seller/services/{id}/archive": {
      "post": {
        "operationId": "SellerServiceController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-services"
        ]
      }
    },
    "/staff/services": {
      "get": {
        "operationId": "StaffServiceController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-services"
        ]
      }
    },
    "/staff/services/{id}": {
      "get": {
        "operationId": "StaffServiceController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-services"
        ]
      }
    },
    "/staff/services/{id}/approve": {
      "post": {
        "operationId": "StaffServiceController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-services"
        ]
      }
    },
    "/staff/services/{id}/reject": {
      "post": {
        "operationId": "StaffServiceController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-services"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bobo&Doda API",
    "description": "Markaziy Osiyo freelance marketplace — milestone escrow, double-entry ledger. Frontend `lib/api` chegarasini qoplaydi.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StaffLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@bobododa.uz"
          },
          "password": {
            "type": "string"
          },
          "totpCode": {
            "type": "string",
            "description": "2FA yoqilgan hisoblar uchun majburiy (TOTP, 6 xona)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "StaffSessionDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "OPERATIONS",
              "FINANCE",
              "SUPPORT",
              "TRUST_SAFETY",
              "KYC_REVIEWER",
              "ADMIN"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "DASHBOARD",
                "USERS",
                "SERVICES",
                "JOBS",
                "ORDERS",
                "KYC",
                "DISPUTES",
                "PAYMENTS",
                "REPORTS",
                "APPEALS",
                "REVIEWS",
                "SUPPORT",
                "CATEGORIES",
                "SETTINGS",
                "AUDIT",
                "STAFF"
              ]
            }
          }
        },
        "required": [
          "accessToken",
          "role",
          "permissions"
        ]
      },
      "StaffMeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "OPERATIONS",
              "FINANCE",
              "SUPPORT",
              "TRUST_SAFETY",
              "KYC_REVIEWER",
              "ADMIN"
            ]
          },
          "title": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "DASHBOARD",
                "USERS",
                "SERVICES",
                "JOBS",
                "ORDERS",
                "KYC",
                "DISPUTES",
                "PAYMENTS",
                "REPORTS",
                "APPEALS",
                "REVIEWS",
                "SUPPORT",
                "CATEGORIES",
                "SETTINGS",
                "AUDIT",
                "STAFF"
              ]
            }
          },
          "mfaEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "role",
          "title",
          "permissions",
          "mfaEnabled"
        ]
      },
...
      "PublicServiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sellerId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "deliveryDays": {
            "type": "number"
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sellerId",
          "categoryId",
          "title",
          "description",
          "lang",
          "price",
          "currency",
          "deliveryDays",
          "createdAt"
        ]
      },
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "price": {
            "type": "number",
            "description": "Butun so‘m",
            "minimum": 1,
            "maximum": 10000000000
          },
          "deliveryDays": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          }
        },
        "required": [
          "categoryId",
          "title",
          "description",
          "price",
          "deliveryDays"
        ]
      },
      "ServiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sellerId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "deliveryDays": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_REVIEW",
              "ACTIVE",
              "REJECTED",
              "PAUSED",
              "ARCHIVED"
            ]
          },
          "submittedAt": {
            "type": "object",
            "nullable": true
          },
          "reviewedAt": {
            "type": "object",
            "nullable": true
          },
          "rejectionReason": {
            "type": "object",
            "nullable": true
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sellerId",
          "categoryId",
          "title",
          "description",
          "lang",
          "price",
          "currency",
          "deliveryDays",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "price": {
            "type": "number",
            "minimum": 1,
            "maximum": 10000000000
          },
          "deliveryDays": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          }
        }
      },
      "RejectServiceDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}